Report TextTypeParser errors as recoverable errors
Previously, `TextTypeParser` would throw exceptions for any errors it
found which meant that parsing would stop at the first error. This
change adds support for `TextTypeParser` to use the new recoverable
error reporting mechanism in `ApiFile` to report the errors. All of its
errors are actually recoverable as `ApiFile` has already extracted the
type from the file and so as long as it returns a valid `TypeItem`
parsing of the rest of the file can continue.

This change adds tests for the different errors and shows that the
parser can continue afterwards. The caching in `TextTypeParser` does
mean that errors are only reported the first time it sees a type
string that is invalid. That negates the main purpose of reporting
recoverable errors which is to report all errors in the file and not
stop at the first one. That will be addressed in a follow-up change.

Bug: 394789173
Test: ./gradlew
Change-Id: I15855b9c2d7c0f58006543b40ddae092b8693b99