Do not assume unqualified type is in java.lang
Previously, the code would assume that any unqualified type in a
signature file was part of the `java.lang` package and so would prefix
it with `java.lang.`. It did that because `java.lang.` was stripped
from types when writing out. The problem with that was that it hid
issues because many unqualified types were not part of `java.lang`.
Those issues were fixed in the previous change.

This change uses reflection to check whether an unqualified type is
part of the `java.lang` package and if it is not reports an issue.

While the `java.lang` package available via reflection is not the
same as the one accessible by API surfaces it should be a super-set. A
follow-up change will extend the `JavaLangPackage` implementation to
check the API surface if possible.

Bug: 394789173
Test: ./gradlew
Change-Id: Ic989442762d2db0381e00f558c10dfccd32d0f9c