Fix a couple of minor issues with Driver.createReportFile(...)
Previously, it would not make sure that the parent directory existed
before creating the `apiFile`. While most callers pass in a command
line option defined using `RawOption.newFile()` which ensures the
parent directory exists, some do not. This change fixes that.

If a parent directory did not exist then this would throw an exception
which would be caught and then reported to `options.reporter`, where
`options` is the global, deprecated variable. However, some of the
callers have disabled access to that in which case an error would
be thrown trying to access it which would hide the original exception.
This change fixes that too by using `codebase.reporter` instead.

Bug: 394597358
Test: ./gradlew
Change-Id: Ieaa4c3c245e272c2b4a50364beeb3e048f0a98c3