Add ability to run multiple input sets of same format
Previously if multiple input sets with the same format were used in a common test, the test would pass with no indication that the test lambda wasn't run for any of the inputs. This runs all of them instead, which is useful for testing signature files of the different formats.

Bug: 377733789
Test: Used in followup changes
Change-Id: I850443581378089599089d0c168a99b3231f3cb4