Add test to show extensions are not sorted last
Originally, `MatchedPatternFile.module` was an indicator of whether
something was from an extension API or a primary API. However, a
specific `extension` property was added to do that but was not added
to the comparator. As a result they are not sorted last which is the
order in which they are processed as extensions are deltas on top of
the primary API.

This change adds a test to show that. A follow-up change will fix it.

Bug: 396431439
Test: ./gradlew
Change-Id: I53bdf48083643447405a87cf44d8545876d8eeb3