Fix test for generating system API from jar file
The previous change added a test for generating the system API from a
jar file using `--api-surface` which caused the consistency check to
fail. This change fixes the consistency check so it is not performed
when the sources are signature files or jar files and fixes the test so
it no longer expects to fail.

It was necessary to baseline the Turbine test as Turbine does not yet
support loading a `Codebase` from a jar file.

Bug: 378479241
Test: ./gradlew
Change-Id: I3e993f29f6f038b0c84d00b86c6135a91bb15a65