Construct ApiSurfaces in AndroidJarsToSignaturesCommand from config
Previously, the `ApiSurfaces` was just created from the names supplied
to `--api-surfaces` option. This change adds support for the
`--config-file` option and creates the `ApiSurfaces` from there instead.
That ensures that it is accurate.

It only includes the `ApiSurface`s that are specified in the
`--api-surfaces` option, or the transitive closure of the surfaces that
they extend. Any other surface in the configuration file is ignored.

Of those `ApiSurface`s it only converts the one supplied to the
`--api-surfaces` option.

So, if `--api-surfaces system` was specified then `ApiSurfaces` will
include both `public` and `system` but it will only convert `system`
files.

This is needed to allow the tool to correctly generate system and other
APIs that are deltas on top of other surfaces.

Bug: 396431439
Test: ./gradlew
Change-Id: I583d03e80741d49665b5f4a5df3ab6cc78071c0e