Refactor SurfaceInfo construction to add extends property
Refactors `SurfaceInfo` construction in `HistoricalApiVersionInfo` to
add an optional `extends` property which references the `SurfaceInfo`
of an extended `ApiSurface`. This is needed to generate a list of all
the historical signature files that contribute to an API surface for a
specific version.

e.g. version 30 of the system API consists of the following two files:
* `prebuilts/sdk/30/public/api/android.txt`
* `prebuilts/sdk/30/system/api/android.txt`

A follow-up change will make use of this.

Bug: 396431439
Test: ./gradlew
Change-Id: I7e0a8a0bbd6b56374366ea1d4b58dc8d47aff0d5