Do not emit libcore.util classes in tests
Does for `libcore.util`s what the preceding changes did for
`android.annotation` and `androidx.annotation`.

The test that previously failed in Turbine did so because the
definition of `libcore.util.NonNull` that was output to the signature
file used the literal value for `Integer.MIN_VALUE` and
`Integer.MAX_VALUE`.

Bug: 398837120
Test: ./gradlew
Change-Id: I99cf7b04ba82ec8eb7de70f5ea8a07c0eaf7975f