Write type aliases to signature files
Any type aliases which are part of the public API surface are now written to signature files. The change in ApiVisitor makes it so a package which has public type aliases but no public classes gets written, and the change in ApiVariantSelectors makes it so a type alias in a hidden package won't be written unless annotated with a show annotation.

Bug: 135191699
Test: SignatureInputOutputTest, HideAnnotationTest and ShowAnnotationTest verify type aliases in hidden packages aren't written unless show-annotated, AndroidX changes: aosp/3727086

Change-Id: I58d06a724fc9b9de027253399563892895437254