Add expect/actual type alias tests
Defining an `expect class` in common with an `actual typealias` in android/jvm is a common pattern in KMP libraries. Currently, usages of these types from common appear as the name of the type alias instead of the value when using K2. This makes sense from the UAST perspective because the value is platform-dependent, but because metalava is just modeling the API for one platform (either android or jvm), it makes more sense to use the value of the type alias.

Bug: 396445075
Test: adds tests
Change-Id: I9fbede888afa8dee36887dc4ffc6fc479d120812