Test @ClassRule in CustomizableParameterizedRunner
Shows the incorrect behavior of `@ClassRule` in
`CustomizableParameterizedRunner`. It should only be applied once but
it is being applied twice, once by `CustomizableParameterizedRunner`
and then again by the `Parameterized` instance to which it delegates.

Following changes will fix that by copying `Parameterized` and
modifying it so it can be extended by `CustomizableParameterizedRunner`
instead of being delegated from it.

Bug: 401196300
Test: ./gradlew
Change-Id: I8037eb0de3751047c83e15c01e25641088b736a3