Wrap default method Anno in an AnnoExpr
Turbine represents a default method value that is an annotation as an
`Anno` (which is not an `Expression`), but all other values are
represented as `Expression`s.

e.g. in the following `@OtherAnno` would be represented as an `Anno`
while `3` would be represented as a `Literal` (which is an
`Expression`):
```
    public @interface MyAnno {
        int number() default 3;
        OtherAnno other() default @OtherAnno;
    }
```

Turbine does have an `AnnoExpr` which wraps an `Anno` in other cases
when it is used as a value (e.g. attribute value) but it is not used in
this case, presumably as a form of optimization.

This change wraps the `Anno` in an `AnnoExpr` as it makes the rest of
the value processing simpler and more consistent.

Bug: 400664604
Test: ./gradlew
Change-Id: I2556b3ed8d230397ee1b8d57236e828a6748eaba