Simplify handling of TurbineSourceFiles
Previously, it was necessary for `createTurbineSourceFile(...)` to be
called to create a `TurbineSourceFile` before it could be retrieved
using `turbineSourceFile(...)`. That was because the latter had no way
to obtain the `CompUnit` object needed by `TurbineSourceFile`. However,
the `classSourceMap` maintained a mapping from `TyDecl` (retrieved from
`SourceTypeBoundClass`) to `CompUnit`. That was used to find the
`CompUnit` to be passed to `createTurbineSourceFile(...)`.

This change combines the two and simplifies that by recognizing that
the `SourceTypeBoundClass` and `CompUnit` both reference `SourceFile`
and `SourceFile`s are uniquely identified by their `path()`.

This passes the list of all `CompUnit`s into `TurbineSourceFileCache`
from which it creates a map from `path` to `CompUnit`. That
necessitated changing it to a `lateinit var`. That map was then used
in `turbineSourceFile(...)` to find the `CompUnit` from which the
`TurbineSourceFile` was created if it could not find an existing
instance.

The `createTurbineSourceFile(...)` method was removed, along with the
call to create `TurbineSourceFile` for `package-info.java` files. The
latter was no longer needed as the `CompUnit`s for packages were
included in the `TurbineSourceFileCache`.

Bug: 400664604
Test: ./gradlew
Change-Id: Id8caaeb91beab4f7d10d93c0e2c3123340c4a341