Detect collisions in the TestFiles in an InputSet
Previously, it was possible to create a set of `TestFile`s for an
`InputSet` whose file names collided. That mean that they would all use
the same `File` and its contents would be the contents of the last
`TestFile` with that name which was created. Changing the nature of the
test from what was intended.

This change adds a check to detect that, fail the test immediately and
fixes the affected tests.

Bug: 402049952
Test: ./gradlew
Change-Id: Ided51857e63f09998c7c5e1d011d540fca10fa9b