Replace check that companion constants are annotated with @JvmField
This lint check finds companion object properties which are val but not const and checks that they are annotated with @JvmField, and not annotated with @JvmStatic. This was previously done through a complicated check on PsiFieldItems which required inspecting the source psi. This replaces it with a much simpler check on the properties themselves.

Bug: 401569415
Test: KotlinInteropChecksTest

Change-Id: Ia04d55fd2dc6dc51ba8fb500a1a128412259d312